Create Password Modifier
Upload of my password modifier program. Many user-created passwords are simple and easy to guess. This program takes a simple password and makes it stronger by replacing characters using the key below, and by appending "!" to the end of the input string.

i becomes 1
a becomes @
m becomes M
B becomes 8
s becomes $

Ex: If the input is:
mypassword

the output is:
Myp@$$word!